# perf(gatsby): include node counts in output after sourcing steps