This will help to triage many sites. If node count is high, certain problems are obvious. Also helps with triaging sites on sources that generate an excessive amount of nodes unexpectedly.

```
success createSchemaCustomization - 0.008s
success source and transform nodes - 29.579s - Now have 100128 nodes with 7 types: [SitePage:1, SitePlugin:22, Site:1, SiteBuildMetadata:1, Directory:1, File:101, IfscCsv:100001]
success building schema - 0.337s
success createPages - 42.945s - Now have 200129 nodes with 7 types: [SitePage:100002, SitePlugin:22, Site:1, SiteBuildMetadata:1, Directory:1, File:101, IfscCsv:100001]
success createPagesStatefully - 0.058s
```