@pvdz 2020-05-01
cc @benrobertsonio 
This should give us more direct insight into perf problems with relatively small sites that manage to generate a ton of internal nodes.