@vladar 2020-05-04
Cool! I think it's quite useful to show the total number of nodes after `source nodes` and `create pages`.

But maybe we should output node count per type in the `--verbose` mode only? There are sites with hundreds of types so these messages will inflate the console and they are really only useful for debugging.