@pvdz 2020-05-04
I would argue that we want to output this by default because that way it will show up in the build output trace that we most often get from people in reports, and the output we see in cloud. This should give you an immediate answer to whether page count / node count is affecting build speeds.