@pvdz 2020-05-06
Ok. Is there a particular reason that there is no `reporter` instance in the bootstrap? Are we recreating a reporter for every step?